+++ core/checks.py
@@ -33,7 +33,7 @@ def has_permissions(permission_level: PermissionLevel = PermissionLevel.REGULAR)
 
         if not has_perm and ctx.command.qualified_name != "help":
             logger.error(
-                "You does not have permission to use this command: `%s` (%s).",
+                "You do not have permission to use this command: `%s` (%s).",
                 str(ctx.command.qualified_name),
                 str(permission_level.name),
             )
@@ -43,7 +43,7 @@ def has_permissions(permission_level: PermissionLevel = PermissionLevel.REGULAR)
     return commands.check(predicate)
 
 
-async def check_permissions(  # pylint: disable=R0911
+async def check_permissions(  # pylint: disable=too-many-return-statements
     ctx, command_name, permission_level
 ) -> bool:
     """Logic for checking permissions for a command for a user"""